requirements.txt
groq
faiss-cpu
sentence-transformers
numpy
streamlit




